refactor packaging structure
